+++ embeddable.config.js
@@ -6,9 +6,11 @@ export default defineConfig({
   lifecycleHooksFile: './lifecycle.config.ts',
   pushModels: false,
 
-  previewBaseUrl: 'https://app.dev.embeddable.com',
-  pushBaseUrl: 'https://api.dev.embeddable.com',
-  audienceUrl: 'https://api.dev.embeddable.com/',
-  authDomain: 'embeddable-dev.eu.auth0.com',
-  authClientId: 'xOKco5ztFCpWn54bJbFkAcT8mV4LLcpG',
+  region: 'EU',
+
+  // previewBaseUrl: 'https://app.dev.embeddable.com',
+  // pushBaseUrl: 'https://api.dev.embeddable.com',
+  // audienceUrl: 'https://api.dev.embeddable.com/',
+  // authDomain: 'embeddable-dev.eu.auth0.com',
+  // authClientId: 'xOKco5ztFCpWn54bJbFkAcT8mV4LLcpG',
 });
